Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,3 +1,4 @@
+// Your code here...
 #include <stdio.h>
 
 int main() {
@@ -5,7 +6,7 @@ int main() {
     char operator;
 
     scanf("%d %d %c", &num1, &num2, &operator);
-
+    
     switch (operator) {
         case '+':
             printf("%d\n", num1 + num2);
@@ -18,7 +19,7 @@ int main() {
             break;
         case '/':
             if (num2 != 0) {
-                printf("%.2f\n", (float)num1 / num2);
+                printf("%d\n", num1 / num2);
             } else {
                 printf("Division by zero\n");
             }
@@ -27,6 +28,6 @@ int main() {
             printf("Invalid operator\n");
             break;
     }
-
+    
     return 0;
 }
